Draw a Hexagon
By following along the tutorial to draw a triangle I was able to
modify it to draw a Hexagon instead. As part of this I also decided
to tidy up the code to make it more extensible by creating classes
for Shaders, BufferObjects, and VertexArrays - these are all very
basic at the moment but can be expanded upon in the future.